Stdlib: openlog/syslog/closelog JIT/AOT via libc helpers (#3676) (#8005)
Add LLVM lowering for the syslog family mirroring VmSyslog auto-openlog
semantics, plus fix StringMktime/__value__writeBool i32 signature drift
that blocked standalone AOT module verification.

+++ lib/JIT/Builtin/StringGmmktime.php
@@ -226,7 +226,7 @@ final class StringGmmktime
             ['gmtime', $i8p, [$i64p]],
             ['timegm', $i64, [$i8p]],
             ['__value__writeLong', $voidTy, [$valuePtr, $i64]],
-            ['__value__writeBool', $voidTy, [$valuePtr, $i8]],
+            ['__value__writeBool', $voidTy, [$valuePtr, $i32]],
         ] as [$name, $ret, $params]) {
             self::ensureExternal(
                 $context,
